Rename version_hash_table.equals/2
Rename version_hash_table.equals/2 to equal/2 as this name is more
consistent with the rest of the standard library.  Normally when renaming
predicates and functions in the standard library we deprecate them.
However, version_hash_table.equals/2 hasn't existed for very long and so
very little code depends on it's name.

library/version_hash_table.m:
    As above.