Avoid C compiler warnings on OS X with Xcode 5.
In the mgnuc script, do not use the name of the C compiler to determine its
type since on OS X with Xcode 5 the executable "gcc" is actually clang.
Instead use the value of C_COMPILER_TYPE determined by the configure script to
set C compiler type, as elsewhere in the Mercury system.

Incorrectly identifying the C compiler type was resulting in lots of
compilation warnings because we were using the warning flags intended for GCC
with clang.  Further warnings were being generated because GCC-only
optimization flags would also be passed to clang.

scripts/mgnuc.in:
	As above.