Fix github issue #14.
Compilation of the Boehm GC was failing on 64-bit openSUSE 13.1 because that
system defines site specific settings for configure scripts and one of those
site specific settings is that when using a 64-bit C compiler, the value of
--libdir should default to ${exec_prefix}/lib64.  For the Boehm GC, this means
that libatomic_ops was being installed libatomic_ops-install/lib64 rather than
libatomic_ops-install/lib.  The top-level Boehm GC Makefile expects to find it
in the latter.

boehm_gc/build_atomic_ops.sh:
	Explicitly set --libdir for libatomic_ops so that we override any
	site specific settings for configure and libatomic_ops is "installed"
 	in the correct directory.