Add discrete interval encoding tree module to the standard library.
Discrete Interval Encoding Trees are a highly efficient set
implementation for fat sets.

This implementation is, in large parts, a translation of Caml Diet.
I have permission from the authors to license this version either under
the LGPL, or LGPL/BSD dual license.  We will choose the LGPL for
consistency with the rest of the standard library, but retain the other
option for the future.

library/diet.m:
	Add the new module.

library/library.m:
	Include the new module in the library.

NEWS:
	Announce the above addition.