Improve README.bootstrap
During the bootstrap process it is necessary to clean the build tree between
the initial and complete builds.  This patch describes this and adds
commands to the instructions that clean the build tree.

As compilation can take a long time I've also added a paragraph describing
how to instruct make to run multiple parallel tasks.

README.bootstrap:
    As above.