Avoid using non-MSVC command line options with MSVC.
The configure script tests the for the presence of various C compiler command
line options by attempting to compile a small test program with those options.
Testing for command line options this way does not work with MSVC since it will
print a warning about the option not being recognised and then carry on as normal.
There is no way to tell MSVC to treat this warning as an error.

The above behaviour resulted in the Mercury.config file for installations
configured to use MSVC containing flags for non-MSVC compilers.

configure.ac:
	Avoid testing for non-MSVC command line options by compiling a test
	program when we are using MSVC as the C compiler.

	Fix a typo: s/boostrap/bootstrap/