Disable executable stripping on Mac OS X for now.
Currently attempting to run dynamically linked executables that are stripped
fails in the presence of programs that use env runtime trace goal conditions.
This is because stripping such executables with just "strip" (as we currently
do) removes the global variable introduced by env runtime trace goal
conditions.  (The long term fix is to invoke strip with the -x flag, but this
needs to be conditional on the value of --mercury-linkage and we currently
don't have the command line options to support that.)

configure.ac:
	Disable stripping of executables on Mac OS X.