Fix the behaviour of string.from_char_list in C# and Java grades.
The predicate string.from_char_list and string.from_reverse_char_list should
throw an exception if their input lists contain the null character.  This
was not done for the C# and Java grades for the former, and the C# grade for
the latter.
(This fixes the failure of tests/hard_coded/null_char in the C# grade and
almost fixes it for the Java grade.)

library/string.m:
	Make the above predicates throw an exception if their input lists
	contain the null character.

NEWS:
	Announce the above fix.