Add the fused multiply-add operation to the math module.
Support the fused multiply-add operation in C grades, where the underlying math
library provides it.  Add a predicate that allows program to test for the
presence of this operation.
(The non-C target languages do not currently provide this operation.)

library/math.m:
	Add the function fma/3.

	Add the predicate have_fma/0 which can be used to test for the
	presence of fma operation in the current grade / platform.

configure.ac:
runtime/mercury_conf.h.in:
	Check whether the C99 fma() function is available.

NEWS:
	Announce the above.