Add boehm_gc and libatomic_ops submodules.
This change adds the boehm_gc and libatomic_ops code as git submodules.
This is analgous to a symbolic link: it's a referece to these external
repositories.

To make things easier, the libatomic_ops submodule is included directly in
Mercury rather than via the boehm_gc submodule.  This makes configuration of
the submodules centralised and a little simpler.

.gitmodules:
boehm_gc:
libatomic_ops:
    As atove.

prepare.sh:
    Add a script that will initialize and checkout the submodules, create a
    symbolic link from boehm_gc/libatomic_ops to libatomic_ops and
    prepare the configure script.