Stand-alone interfaces and non-C grades.
Clarify the situation with stand-alone interfaces and non-C grades.  The
documentation currently says that they are not currently supported for the
non-C grades, which is inaccurate.  They are not actually required for the Java
and C# backends.  The may however be needed for Erlang.

doc/user_guide.texi:
	Stand-alone interface are *not* required if the target language is
	C# or Java.

compiler/mercury_compiler.m:
	Fix an XXX: generate a more accurate error message if
	'--generate-standalone-interface' is used with one of the
	non-C target languages.