Get configure to repeat some important parameters.
Some of the most important autoconfigured parameters are echoed by the
configure script in the middle of its long output, making them hard to find.
Repeat these parameters (the installed compiler if any, the default grade,
the grade the workspace will be compiled in, and the target prefix if any)
at the end, to make them easy to see without scrolling back or searching.

Fix some other minor nits.