Improvements to the reference manual and user's guide.
This change addresses three issues in the reference manual and user's guide
that mainly affect the appearance of the PDF versions of those documents.
These three issues are:

1. The indentation of many of the examples is broken in the PDF version due
to the texinfo source for those examples containing tabs instead of spaces.
Moreover, many of the examples are also over-indented and end up running over
the RHS margin.  (These problems don't really show up in the HTML version.)

2. There are many awkward line and page breaks in the PDF version, for instance
in the middle of examples.  This change adds @group and @w commands in order to
prevent this from happening.  It also forces chapters to begin on a new page,
which assists with the breaking issue (and also looks less weird than starting
a new chapter in middle of a page.)

3. There is a lot of inconsistency over the use of @samp, @code and @var.
Whilst the distinction between the first two is sometimes just a matter
of taste, our existing usage is just all over the place.

Fix a bunch of other minor issues (noted below).

doc/reference_manual.texi:
	Fix a typo in the field update example.

	Fix foreign_proc examples that would result in a compilation error:
	many of them are missing a 'promise_pure' attribute.

	s/typeclass constraint/type class constraint/

	Re-format foreign_procs as per the Mercury project coding standard.
	This prevents them overrunning the RHS margin in many cases.

	The bibliography lists a paper by Demoen and Sagonas as "submitted
	for publication".  Replace that with the actual publication details.

doc/user_guide.texi:
	Add missing texinfo markup.

compiler/options.m:
	Delete a couple of extraneous tabs in the usage message.