Fix handling of newlines in configure tests.
When using MSVC as the C compiler on Cygwin, configuration test programs
(for example, the one that determines the C type of 64-bit integer)
write out CRLF line endings.  However, the shell in Cygwin may be
configured to only treat LF as a line ending.  In that case the CR
character will be treated as part of the word and that creates problems
like the following:

    checking for an integer type with the same size as a pointer... int
    checking for an integer type of at least 64 bits... __int64 configure:
    error: Cannot determine the length modifier for the MR_int_least64_t
    type.

Configure can't match __int64^M against __int64 and therefore cannot
determine the correct length modifier

This patch removes the "\n" escapes from the relevant printf calls.

configure.ac:
    As above.