Handle exit codes of forked processes on windows safely.
MSVC doesn't provide the WIFEXITED and other macros that help a programmer
determine if a child process exited with an error message or was killed by a
signal.  In these cases Mercury assumes that positive numbers are exit codes
and negative numbers mean that a process was killed by a signal.

When "cl" exited with errors, returning the error code -2, mmc --make
thought that it was killed by a signal and failed to display the compilation
errors.

This patch removes the existing code that tries to detect signals when these
macros are not available.  As far as I know platforms like windows don't
have a concept of SysV signals, therefore signals can only occur when using
cygwin or some other compatibility layer which will then provide the
WIFEXITED and other macros.

I considered using the windows API directly however it didn't (as far as I
could find) provide any concept of signals, so I rejected this idea.

library/io.m:
    As above.

    Also avoid the use of magic numbers such as 127 for failure.

    Rename the handle_system_command_exit_status/1 function to
    decode_system_command_exit_code,

compiler/process_util.m:
    Conform to changes in io.m

compiler/module_cmds.m:
    If a subprocess is killed by a signal then report this both to the
    desired stream (the error stream for the subtask) and to standard
    output.

    After reporting the error then re-raise the signal, not before.

compiler/file_util.m:
    Flush output after reporting an error to make sure that it is seen.