Fix problems with --use-grade-subdirs with C# and Java backends.
Fix problems that occur when generating executables in the C# and Java grades
with --use-grade-subdirs enabled, namely:

* the compiler aborts when it tries to build the launcher script for a Java
grade executable.
* the compiler does not place generated Java archive files under the Mercury
subdirectory.
* in C# grades, the compiler always places the launcher script in the user's
directory rather than under the Mercury subdirectory.

The above problems were mainly caused by the "link" step for executables being
incorrect for the C# and Java backends.  In other grades, the link step
consists of:

1. use the linker to create the executable in C grades, or generate a launcher
script in the Erlang grade.  This executable or script is placed under the
Mercury subdirectory.

2. create a symlink (or copy if --no-use-symlinks is enabled) the executable
or script into the user's directory.

This does not work for the C# and Java backends since their principle "link
target" is not the launcher script (as is the case with the Erlang backend),
but a ".exe" or ".jar" file.  The launcher script is generated as a side-effect
of creating that.   (This has long been the case for the C# backend and has
recently become the case for the Java backend since we moved to generating
archives for executables.)  The correct sequence of "linking" steps for these
two backends should be:

1. create the .exe or .jar under the Mercury subdirectory.
2. create the launcher script file, if required, under the Mercury subdirectory.
3. symlink or copy the .exe or .jar file to the user's directory.
4. symlink or copy the launcher script to the user's directory, if required.

This diff makes the compiler carry out the latter sequence of steps for the C#
and Java backends.

compiler/compile_target_code.m:
	Make the executable target for the Java backend work more like that of
	the C# backend.  Merge it with the code for handling library targets in the
	Java backend since, apart from the creation of the launcher script, it is
	identical.

	We may need to do multiple create symlink or copy operations when
	creating an executable using the C# or Java backends: one for the
	".exe" or ".jar" and another for the launcher script.

	Separate out the code for determining the launcher script extension into
	a new function.

compiler/make.util.m:
	Conform to the above change.

	Add an XXX comment about the Erlang backend; I'll look into this
	separately.

compiler/file_names.m:
	Update the list of architecture or grade dependent files to include those
	with the extensions ".jar" and ".bat", and also those files with no
	extension at all.  (For Mercury, the latter will either be launcher scripts
	or executables.)   Not covering these cases had a number of untoward
	effects: ".jar" files were not placed in grade specific directories and the
	launcher script in the C# grade was always placed in the user's directory.

compiler/module_cmds.m:
	Create any required subdirectories under the Mercury directory when
	generating launcher scripts.

	If --verbose-commands is enabled, then print a message when creating
	a symlink or copying a file.  I needed this when debugging the above
	problems.

	Use io.format/4 to generate an error message rather than multiple
	calls to io.write_string/3.